+++ lib/webspicy/web/openapi.rb
@@ -1,1 +1,2 @@
 require_relative 'openapi/generator'
+require_relative 'openapi/ext'