+++ lib/webspicy/web/openapi/ext.rb
@@ -1,0 +1,18 @@
+module Webspicy
+  class Specification
+    module Pre
+
+      def contribute_to_openapi_verb(verb_defn)
+        verb_defn
+      end
+
+    end # module Pre
+    module Post
+
+      def contribute_to_openapi_verb(verb_defn)
+        verb_defn
+      end
+
+    end # module Post
+  end # class Specification
+end # module Webspicy